+++ src/lib.rs
@@ -263,7 +263,10 @@ mod de;
 mod error;
 mod macros;
 mod ser;
-pub mod set;
+
+pub mod binary_set;
+pub mod number_set;
+pub mod string_set;
 
 pub use attribute_value::{AttributeValue, Item, Items};
 pub use de::{from_attribute_value, from_item, from_items, Deserializer};
